refactor(log-context): move LoggingContext to Core\Log namespace
Phpstorm refactor
